Move alternative money formatters to a separate Gradle module
