+++ kairo-alternative-money-formatters/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+  id("kairo")
+  id("kairo-publish")
+}
+
+dependencies {
+  implementation(project(":kairo-money"))
+  implementation(project(":kairo-serialization"))
+
+  testImplementation(project(":kairo-testing"))
+}
