+++ kairo-serialization/src/main/kotlin/kairo/serialization/module/money/MoneyFormatter.kt
@@ -9,6 +9,28 @@ import kotlin.reflect.KClass
 import org.javamoney.moneta.Money
 
 public abstract class MoneyFormatter<T : Any> {
+  /**
+   * This is a good standard format to serialize money.
+   * It includes currency information and does not lose precision.
+   */
+  public object Default : MoneyFormatter<Default.Intermediary>() {
+    public data class Intermediary(
+      val amount: String,
+      val currency: String,
+    )
+
+    override fun parse(value: Any): Money {
+      value as Intermediary
+      return Money.of(BigDecimal(value.amount), value.currency)
+    }
+
+    override fun format(money: Money): Intermediary =
+      Intermediary(
+        amount = amountAsString(money),
+        currency = currencyCode(money),
+      )
+  }
+
   internal val kClass: KClass<T> = typeParam(MoneyFormatter::class, 0, this::class)
 
   public abstract fun parse(value: Any): Money
@@ -26,7 +48,7 @@ public abstract class MoneyFormatter<T : Any> {
 }
 
 public var ObjectMapperFactory<*, *>.moneyFormatter: MoneyFormatter<*>
-  get() = properties["moneyFormatter"] as MoneyFormatter<*>? ?: DefaultMoneyFormatter
+  get() = properties["moneyFormatter"] as MoneyFormatter<*>? ?: MoneyFormatter.Default
   set(value) {
     properties["moneyFormatter"] = value
   }
